tests: share native crash handler between Graphics.Regression and AutoTesting
Extract SetErrorMode + FirstChanceException minidump capture into sources/shared/NativeCrashHandler.cs, used by both. Also fixes Graphics.Regression setting SEM_NOGPFAULTERRORBOX unconditionally, which blocked dump capture for pure-native crashes; the shared handler gates it off in capture mode.
