+++ sources/engine/Stride.Games.AutoTesting/Stride.Games.AutoTesting.csproj
@@ -11,6 +11,9 @@
     <Compile Include="..\..\shared\SharedAssemblyInfo.cs">
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
+    <Compile Include="..\..\shared\NativeCrashHandler.cs">
+      <Link>NativeCrashHandler.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Stride.Engine\Stride.Engine.csproj" />
