+++ sources/engine/Stride.Graphics.Regression/Stride.Graphics.Regression.csproj
@@ -16,6 +16,9 @@
     <Compile Include="..\..\shared\SharedAssemblyInfo.cs">
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
+    <Compile Include="..\..\shared\NativeCrashHandler.cs">
+      <Link>NativeCrashHandler.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="xunit" />
